+++ plone/formwidget/datetime/at/tests/test_MonthYearWidget.py
@@ -20,7 +20,6 @@ class TestMonthYearWidget(unittest.TestCase):
             {
                 'show_calendar': True,
                 'helper_css': (),
-                'with_time': False,
                 'years_range': (-10, 10),
                 'description': '',
                 'populate': True,