+++ plone/formwidget/datetime/tests/test_AbstractDatetimeWidget.py
@@ -59,13 +59,13 @@ class TestAbstractDatetimeWidget(unittest.TestCase):
         instance.request.get.return_value = '55'
         self.assertEqual(instance.minute, '55')
 
-    def test_minute_is_None_value_not_empty_value(self):
+    def test_aminute_is_None_value_not_empty_value(self):
         instance = self.createInstance()
         instance.name = 'field'
         instance.request = {}
         instance.value = '12345'
         instance.empty_value = '45678'
-        self.assertEqual(instance.minute, '8')
+        self.assertEqual(instance.minute, '5')
 
     def test_minute_is_None_value_is_empty_value(self):
         instance = self.createInstance()