+++ plone/formwidget/datetime/z3cform/converter.py
@@ -78,6 +78,8 @@ class MonthYearDataConverter(DateDataConverter):
             return ('', '', '1')
         return (value.year, value.month, value.day)
 
+    def toFieldValue(self, value):
+        import pdb;pdb.set_trace()  ## Breakpoint ##
 
 class YearDataConverter(DateDataConverter):